Split DataPointConversion.kt comment, merge, and operand helpers into separate files
